cleanup and resort the current code for the update service

+++ Helper.Tools.Updater/ViewModels/UpdateCheckerViewModel.cs
@@ -11,7 +11,7 @@
     using Properties;
 
     /// <summary>The UpdateCheckerViewModel.</summary>
-    public class UpdateCheckerViewModel : ViewModelBase, IUpdateable
+    public class UpdateCheckerViewModel : ViewModelBase
     {
         #region Fields
         #endregion Fields
@@ -31,24 +31,11 @@
 
         /// <summary>Gets or sets the current version.</summary>
         /// <value>The current version.</value>
-        public string CurrentVersion { get; set; } = Assembly.GetEntryAssembly().GetName().Version.ToString();
+        public string CurrentVersion => Assembly.GetEntryAssembly().GetName().Version.ToString();
 
         /// <summary>Gets or sets the last change.</summary>
         /// <value>The last change.</value>
-        public string LastChange { get; set; }
-
-        #region Implementation of IUpdateable
-        /// <summary>The name of your application as you want it displayed on the update form.</summary>
-        public string ApplicationName => "TestApp";
-
-        /// <summary>An identifier string to use to identify your application in the update.json.</summary>
-        /// <remarks>Should be the same as your appId in the update.json.</remarks>
-        public string ApplicationId => "TestApp";
-
-        /// <summary>The location of the update.json on a server.</summary>
-        public Uri UpdateJsonLocation => new Uri("");
-        #endregion
-
+        public string LastChange { get; private set; }
         #endregion Properties
 
         #region Methods
@@ -58,15 +45,22 @@
             if (!Directory.Exists(tmpPath)) Directory.CreateDirectory(tmpPath);
 
             var tmpFile = tmpPath + "version.txt";
-            new WebClient().DownloadFile(Settings.Default.WebUpdateVersionFile, tmpFile);
+
+            if (!File.Exists(tmpFile) || (File.GetCreationTimeUtc(tmpFile).Date - DateTime.Now).TotalHours < 2)
+            {
+                if (!WebService.ExistsOnServer(new Uri(Settings.Default.WebUpdateVersionFile))) return;
+
+                new WebClient().DownloadFile(Settings.Default.WebUpdateVersionFile, tmpFile);
+            }
 
             NewVersion = File.ReadLines(tmpFile).First();
-
             LastChange = string.Empty;
 
             foreach (var line in FileService.Read(tmpFile))
             {
-                if (line.StartsWith("[")) LastChange += line + "\n";
+                if (!line.StartsWith("[")) continue;
+
+                LastChange += line.Replace("<br />", string.Empty) + "\n";
             }
 
             Directory.Delete(tmpPath, true);
